[MemoryPressure] Handle video memory usage besides normal memory.

+++ Source/WTF/wtf/MemoryPressureHandler.cpp
@@ -26,6 +26,7 @@
 #include "config.h"
 #include <wtf/MemoryPressureHandler.h>
 
+#include <fnmatch.h>
 #include <wtf/Logging.h>
 #include <wtf/MemoryFootprint.h>
 #include <wtf/NeverDestroyed.h>
@@ -44,6 +45,55 @@ static const double s_strictThresholdFraction = 0.5;
 #endif
 static const std::optional<double> s_killThresholdFraction;
 static const Seconds s_pollInterval = 30_s;
+
+// This file contains the amount of video memory used, and will be filled by some other
+// platform component. It's a text file containing an unsigned integer value.
+static String s_GPUMemoryFile;
+
+static bool isWebProcess()
+{
+    static bool result = false;
+    static bool initialized = false;
+
+    if (!initialized) {
+        initialized = true;
+
+        FILE* file = fopen("/proc/self/cmdline", "r");
+        if (!file)
+            return result;
+
+        char* buffer = nullptr;
+        size_t size = 0;
+        if (getline(&buffer, &size, file) != -1)
+            result = !fnmatch("*WPEWebProcess*", buffer, 0);
+
+        free(buffer);
+        fclose(file);
+    }
+
+    return result;
+}
+
+static size_t memoryFootprintVideo()
+{
+    if (!isWebProcess() || s_GPUMemoryFile.isEmpty())
+        return 0;
+
+    FILE* file = fopen(s_GPUMemoryFile.utf8().data(), "r");
+    if (!file)
+        return 0;
+
+    char* buffer = nullptr;
+    size_t size = 0;
+    size_t footprint = 0;
+    if (getline(&buffer, &size, file) != -1)
+        sscanf(buffer, "%u", &footprint);
+
+    free(buffer);
+    fclose(file);
+
+    return footprint;
+}
 
 MemoryPressureHandler& MemoryPressureHandler::singleton()
 {
@@ -65,6 +115,11 @@ MemoryPressureHandler::MemoryPressureHandler()
 #if PLATFORM(COCOA)
     setDispatchQueue(dispatch_get_main_queue());
 #endif
+
+    // If this is the WebProcess, Check whether the env var WPE_POLL_MAX_MEMORY_GPU_FILE exists, containing the file
+    // that we need to poll to get the video memory used.
+    if (isWebProcess())
+        s_GPUMemoryFile = String::fromLatin1(getenv("WPE_POLL_MAX_MEMORY_GPU_FILE"));
 }
 
 void MemoryPressureHandler::setShouldUsePeriodicMemoryMonitor(bool use)
@@ -123,10 +178,14 @@ void MemoryPressureHandler::setPageCount(unsigned pageCount)
     singleton().m_pageCount = pageCount;
 }
 
-std::optional<size_t> MemoryPressureHandler::thresholdForMemoryKill()
+std::optional<size_t> MemoryPressureHandler::thresholdForMemoryKill(MemoryType type)
 {
     if (m_configuration.killThresholdFraction)
-        return m_configuration.baseThreshold * (*m_configuration.killThresholdFraction);
+        return (*m_configuration.killThresholdFraction) * (type == MemoryType::Normal ? m_configuration.baseThreshold : m_configuration.baseThresholdVideo);
+    else {
+        // Don't kill the process if no killThreshold was set.
+        return std::nullopt;
+    }
 
     switch (m_processState) {
     case WebsamProcessState::Inactive:
@@ -137,57 +196,61 @@ std::optional<size_t> MemoryPressureHandler::thresholdForMemoryKill()
     return std::nullopt;
 }
 
-size_t MemoryPressureHandler::thresholdForPolicy(MemoryUsagePolicy policy)
+size_t MemoryPressureHandler::thresholdForPolicy(MemoryUsagePolicy policy, MemoryType type)
 {
     switch (policy) {
     case MemoryUsagePolicy::Unrestricted:
         return 0;
     case MemoryUsagePolicy::Conservative:
-        return m_configuration.baseThreshold * m_configuration.conservativeThresholdFraction;
+        return m_configuration.conservativeThresholdFraction * (type == MemoryType::Normal ? m_configuration.baseThreshold : m_configuration.baseThresholdVideo);
     case MemoryUsagePolicy::Strict:
-        return m_configuration.baseThreshold * m_configuration.strictThresholdFraction;
+        return m_configuration.strictThresholdFraction * (type == MemoryType::Normal ? m_configuration.baseThreshold : m_configuration.baseThresholdVideo);
     default:
         ASSERT_NOT_REACHED();
         return 0;
     }
 }
 
-MemoryUsagePolicy MemoryPressureHandler::policyForFootprint(size_t footprint)
+MemoryUsagePolicy MemoryPressureHandler::policyForFootprints(size_t footprint, size_t footprintVideo)
 {
-    if (footprint >= thresholdForPolicy(MemoryUsagePolicy::Strict))
+    if (footprint >= thresholdForPolicy(MemoryUsagePolicy::Strict, MemoryType::Normal) || footprintVideo >= thresholdForPolicy(MemoryUsagePolicy::Strict, MemoryType::Video))
         return MemoryUsagePolicy::Strict;
-    if (footprint >= thresholdForPolicy(MemoryUsagePolicy::Conservative))
+    if (footprint >= thresholdForPolicy(MemoryUsagePolicy::Conservative, MemoryType::Normal) || footprintVideo >= thresholdForPolicy(MemoryUsagePolicy::Conservative, MemoryType::Video))
         return MemoryUsagePolicy::Conservative;
     return MemoryUsagePolicy::Unrestricted;
 }
 
 MemoryUsagePolicy MemoryPressureHandler::currentMemoryUsagePolicy()
 {
-    return policyForFootprint(memoryFootprint());
+    return policyForFootprints(memoryFootprint(), memoryFootprintVideo());
 }
 
-void MemoryPressureHandler::shrinkOrDie(size_t killThreshold)
+void MemoryPressureHandler::shrinkOrDie(size_t killThreshold, size_t killThresholdVideo)
 {
     RELEASE_LOG(MemoryPressure, "Process is above the memory kill threshold. Trying to shrink down.");
     releaseMemory(Critical::Yes, Synchronous::Yes);
 
     size_t footprint = memoryFootprint();
+    size_t footprintVideo = memoryFootprintVideo();
     RELEASE_LOG(MemoryPressure, "New memory footprint: %zu MB", footprint / MB);
 
-    if (footprint < killThreshold) {
+    if ((footprint < killThreshold) && (footprintVideo < killThresholdVideo)) {
         RELEASE_LOG(MemoryPressure, "Shrank below memory kill threshold. Process gets to live.");
-        setMemoryUsagePolicyBasedOnFootprint(footprint);
+        setMemoryUsagePolicyBasedOnFootprints(footprint, footprintVideo);
         return;
     }
 
-    WTFLogAlways("Unable to shrink memory footprint of process (%zu MB) below the kill thresold (%zu MB). Killed\n", footprint / MB, killThreshold / MB);
+    if (footprint >= killThreshold)
+        WTFLogAlways("Unable to shrink memory footprint of process (%zu MB) below the kill thresold (%zu MB). Killed\n", footprint / MB, killThreshold / MB);
+    else
+        WTFLogAlways("Unable to shrink video memory footprint of process (%zu MB) below the kill thresold (%zu MB). Killed\n", footprintVideo / MB, killThresholdVideo / MB);
     RELEASE_ASSERT(m_memoryKillCallback);
     m_memoryKillCallback();
 }
 
-void MemoryPressureHandler::setMemoryUsagePolicyBasedOnFootprint(size_t footprint)
+void MemoryPressureHandler::setMemoryUsagePolicyBasedOnFootprints(size_t footprint, size_t footprintVideo)
 {
-    auto newPolicy = policyForFootprint(footprint);
+    auto newPolicy = policyForFootprints(footprint, footprintVideo);
     if (newPolicy == m_memoryUsagePolicy)
         return;
 
@@ -199,16 +262,18 @@ void MemoryPressureHandler::setMemoryUsagePolicyBasedOnFootprint(size_t footprin
 void MemoryPressureHandler::measurementTimerFired()
 {
     size_t footprint = memoryFootprint();
+    size_t footprintVideo = memoryFootprintVideo();
 #if PLATFORM(COCOA)
     RELEASE_LOG(MemoryPressure, "Current memory footprint: %zu MB", footprint / MB);
 #endif
-    auto killThreshold = thresholdForMemoryKill();
-    if (killThreshold && footprint >= *killThreshold) {
-        shrinkOrDie(*killThreshold);
+    auto killThreshold = thresholdForMemoryKill(MemoryType::Normal);
+    auto killThresholdVideo = thresholdForMemoryKill(MemoryType::Video);
+    if ((killThreshold && footprint >= *killThreshold) || (killThresholdVideo && footprintVideo >= *killThresholdVideo)) {
+        shrinkOrDie(*killThreshold, *killThresholdVideo);
         return;
     }
 
-    setMemoryUsagePolicyBasedOnFootprint(footprint);
+    setMemoryUsagePolicyBasedOnFootprints(footprint, footprintVideo);
 
     switch (m_memoryUsagePolicy) {
     case MemoryUsagePolicy::Unrestricted:
@@ -327,6 +392,7 @@ void MemoryPressureHandler::setDispatchQueue(OSObjectPtr<dispatch_queue_t>&& que
 
 MemoryPressureHandler::Configuration::Configuration()
     : baseThreshold(std::min(3 * GB, ramSize()))
+    , baseThresholdVideo(1 * GB)
     , conservativeThresholdFraction(s_conservativeThresholdFraction)
     , strictThresholdFraction(s_strictThresholdFraction)
     , killThresholdFraction(s_killThresholdFraction)
@@ -334,8 +400,9 @@ MemoryPressureHandler::Configuration::Configuration()
 {
 }
 
-MemoryPressureHandler::Configuration::Configuration(size_t base, double conservative, double strict, std::optional<double> kill, Seconds interval)
+MemoryPressureHandler::Configuration::Configuration(size_t base, size_t baseVideo, double conservative, double strict, std::optional<double> kill, Seconds interval)
     : baseThreshold(base)
+    , baseThresholdVideo(baseVideo)
     , conservativeThresholdFraction(conservative)
     , strictThresholdFraction(strict)
     , killThresholdFraction(kill)
